i dislike C(++)

+++ src/fft.hh
@@ -1,8 +1,6 @@
 #pragma once
 
-// Source - https://stackoverflow.com/a/37729648
-// Posted by Leos313, modified by community. See post 'Timeline' for change history
-// Retrieved 2026-04-25, License - CC BY-SA 4.0
+// Source - https://it.wikipedia.org/wiki/Trasformata_di_Fourier_veloce
 
 #include <iostream>
 #include <complex>
@@ -35,20 +33,20 @@ int reverse(int N, int n)    //calculating revers number
   return p;
 }
 
-void ordina(complex<double>* f1, int N) //using the reverse order in the array
+void ordina(complex<float>* f1, int N) //using the reverse order in the array
 {
-  complex<double> f2[MAX];
+  complex<float> f2[MAX];
   for(int i = 0; i < N; i++)
     f2[i] = f1[reverse(N, i)];
   for(int j = 0; j < N; j++)
     f1[j] = f2[j];
 }
 
-void transform(complex<double>* f, int N) //
+void transform(complex<float>* f, int N) //
 {
   ordina(f, N);    //first: reverse order
-  complex<double> *W;
-  W = (complex<double> *)malloc(N / 2 * sizeof(complex<double>));
+  complex<float> *W;
+  W = (complex<float> *)malloc(N / 2 * sizeof(complex<float>));
   W[1] = polar(1., -2. * M_PI / N);
   W[0] = 1;
   for(int i = 2; i < N / 2; i++)
@@ -58,8 +56,8 @@ void transform(complex<double>* f, int N) //
   for(int j = 0; j < log2(N); j++) {
     for(int i = 0; i < N; i++) {
       if(!(i & n)) {
-        complex<double> temp = f[i];
-        complex<double> Temp = W[(i * a) % (n * a)] * f[i + n];
+        complex<float> temp = f[i];
+        complex<float> Temp = W[(i * a) % (n * a)] * f[i + n];
         f[i] = temp + Temp;
         f[i + n] = temp - Temp;
       }
@@ -70,7 +68,7 @@ void transform(complex<double>* f, int N) //
   free(W);
 }
 
-void FFT(complex<double>* f, int N, double d)
+void FFT(complex<float>* f, int N, float d)
 {
   transform(f, N);
   for(int i = 0; i < N; i++)
